JL8-106 refactor : HistoryListNone컴포넌트 분리
