+++ src/app/music-history/_components/HistoryList.tsx
@@ -1,15 +1,13 @@
 'use client'
 import PlaylistItem from '@/app/playlist/_components/PlaylistItem'
-import { BorderButton } from '@/components/common'
 import { useMusicPlayerStore } from '@/store/useMusicPlayerStore'
 import { userStore } from '@/store/userSlice'
 import type { Tables } from '@/types/supabase'
 import clsx from 'clsx'
-import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
+import HistoryListNone from './HistoryListNone'
 
 const HistoryList = () => {
-  const router = useRouter()
   const { user } = userStore()
   const { isPlayerOpen, setTrackIds, setPlayerOpen, play } =
     useMusicPlayerStore()
@@ -75,26 +73,7 @@ const HistoryList = () => {
               />
             ))
           ) : (
-            <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2">
-              <div
-                className={clsx(
-                  'flex w-[226px] flex-col items-center justify-center gap-3',
-                )}
-              >
-                <h1 className="title-2">아직 재생목록이 없습니다.</h1>
-                <p className="caption-2 text-center text-opacity-40">
-                  지금 Zepplaylist에서 인기있는 곡을 듣고
-                  <br />
-                  재생목록을 만들어보세요
-                </p>
-                <BorderButton
-                  onClick={() => router.push('/koreaTopChart')}
-                  className="px-[18px] py-[11px]"
-                >
-                  TOP100 바로가기
-                </BorderButton>
-              </div>
-            </div>
+            <HistoryListNone />
           )}
         </ul>
       </div>
